requirements.txt
absl-py==2.1.0
alabaster==0.7.16
albumentations==1.3.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appnope==0.1.4
arabic-reshaper==3.0.0
asn1crypto==1.5.1
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
boto3==1.34.87
botocore==1.34.87
build==1.2.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
coverage==5.3.1
cryptography==42.0.5
cssselect2==0.7.0
cycler==0.11.0
data-gradients==0.3.2
decorator==5.1.1
Deprecated==1.2.14
docutils==0.17.1
einops==0.3.2
exceptiongroup==1.2.0
fastapi==0.103.2
flatbuffers==24.3.25
fonttools==4.51.0
future==1.0.0
grpcio==1.62.2
h11==0.14.0
html5lib==1.1
humanfriendly==10.0
hydra-core==1.3.2
idna==3.6
imagededup==0.3.2
imageio==2.34.0
imagesize==1.4.1
importlib-metadata==6.7.0
ipython==7.34.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.0
json-tricks==3.16.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
numpy==1.23.0
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.0
onnxsim==0.4.36
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
oscrypto==1.3.0
packaging==23.2
pandas==1.3.5
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pip-tools==7.4.1
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pycparser==2.22
pydantic==2.5.3
pydantic_core==2.14.6
pyDeprecate==0.3.2
Pygments==2.17.2
pyHanko==0.23.2
pyhanko-certvalidator==0.26.3
pyparsing==3.1.1
pypdf==4.2.0
pypng==0.20220715.0
pyproject_hooks==1.0.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
qrcode==7.4.2
qudida==0.0.4
rapidfuzz==3.8.1
referencing==0.34.0
reportlab==3.6.13
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
s3transfer==0.10.1
scikit-image==0.22.0
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.12.2
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
Sphinx==4.0.3
sphinx-rtd-theme==1.3.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
starlette==0.27.0
stringcase==1.2.0
super-gradients==3.7.1
svglib==1.5.1
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==1.1.0
threadpoolctl==3.4.0
tifffile==2024.4.18
tinycss2==1.2.1
tomli==2.0.1
torch==1.13.1
torchmetrics==0.8.0
torchvision==0.14.1
tqdm==4.66.2
traitlets==5.9.0
treelib==1.6.1
typing_extensions==4.7.1
tzlocal==5.2
uritools==4.0.2
urllib3==2.0.7
uvicorn==0.22.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.2
wrapt==1.16.0
xhtml2pdf==0.2.11
zipp==3.15.0

